taskmanager/requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.1
django-extensions==4.1
django-filter==24.3
django-simple-history==3.7.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
exceptiongroup==1.2.2
gunicorn==23.0.0
icdiff==2.0.7
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mypy_extensions==1.1.0
openapi-codec==1.3.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pprintpp==0.4.0
psycopg2-binary==2.9.9
pytest==8.3.3
pytest-django==4.9.0
pytest-icdiff==0.9
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
ruff==0.11.7
simplejson==3.20.1
sqlparse==0.5.1
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.4.0
